requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
asttokens==2.4.1
    # via stack-data
build==1.2.2
    # via pip-tools
click==8.1.7
    # via pip-tools
decorator==5.1.1
    # via
    #   ipdb
    #   ipython
executing==2.1.0
    # via stack-data
filelock==3.16.1
    # via torch
fsspec==2024.9.0
    # via torch
ipdb==0.13.13
    # via -r requirements.in
ipython==8.28.0
    # via ipdb
jedi==0.19.1
    # via ipython
jinja2==3.1.4
    # via torch
markupsafe==2.1.5
    # via jinja2
matplotlib-inline==0.1.7
    # via ipython
mpmath==1.3.0
    # via sympy
networkx==3.3
    # via torch
numpy==2.1.1
    # via -r requirements.in
packaging==24.1
    # via build
parso==0.8.4
    # via jedi
pexpect==4.9.0
    # via ipython
pip-tools==7.4.1
    # via -r requirements.in
prompt-toolkit==3.0.48
    # via ipython
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.3
    # via stack-data
pygments==2.18.0
    # via ipython
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
python-mnist==0.7
    # via -r requirements.in
six==1.16.0
    # via asttokens
stack-data==0.6.3
    # via ipython
sympy==1.13.3
    # via torch
torch==2.4.1
    # via -r requirements.in
traitlets==5.14.3
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.12.2
    # via torch
wcwidth==0.2.13
    # via prompt-toolkit
wheel==0.44.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
